Adds shuffled condition

This function add a condition variable that contains random condition labels to a dataframe.

shuffle_cond(d, ni, np, ntc)

Arguments

d
a dataset to which a shuffled condition variable will be added

ni
the number of iterations included in the dataframe

np
the number of participants included in the dataframe

ntc
number of trials per condition

Details

The condition labels will be c('A', 'B'). As it is used in a simulation environment this labels are arbitrary.

It assumes:

The same number of participants across iterations

The same number of trials across all participants

The same number of trials per condition

That the dataframe is sorted by iteration and participant (It would be enough for each participant to be in consecutive rows)

An even number of trials in the whole participant. (Must be if it has two conditions and the same number of trials per condition)#'

Examples

 shuffle_cond(d, ni, np, ntc)

#> Error in lapply(1:np * ni, function(x) {    sample(rep(c("A", "B"), ntc))}): object 'np' not found